Update packages links in qiskit-devs
Currently there is one link named `proper packages` that does not have a
values specified for it. There is also another link named packages which
points (I'm guessing here) to an IBM internal repository.

This commit suggests updating both links to point to the packages
directory. My reasoning here is that these should point to the currently
supported packages, qiskit-devs-anu, qiskit-devs-ibm, and
qiskit-devs-js.