Rename qbts to unusedWires in createTransform
This commit renames the qbts variable in Circuit.createTransform to
unusedWires in hopes to make it more understandable.

My understanding is that this array represents the unused wires
(indexed in opposite order from the qubits/wires passed into the
createTransform function).

For example, a system of 3 qubits and a single hadamard gate:

qubits                                opposite-order
 q[0] ----H--------------------------- q[2]
 q[1] -------------------------------- q[1]
 q[2] -------------------------------- q[0]

In this case qubits/wires used by this gate would be [2] and
qbts/unsuedWires would be [0, 1] (using the opposite-order indexing).