Remove clear() call from Circuit constructor
This commit removes the clear call from the Circuit constructor and puts
the member initialisation of gates, state, and T, in the constructor
itself.

The motivation for this is to hopefully clarify the state of a
Circuit making it a little easier to follow when one does not have to
follow the clear() -> reset() -> resetTransform() calls to figure this
out.