Allow for executing the same circuit twice
This commit adds support for executing the same circuit twice (or more
times) in row with the state from the previous execution being used as
the state for the next execution.

The motivation for this came out of wanting to executing a hadamard gate
twice, and while it is possible to add two gates and then run it, having
the ability to run the circuit multiple times gives a nice separation
for someone learning about quantum gates (allows for inspecting/logging
the state if one is not comfortable using a debugger for example).