Add print circuit method
This commit introduces a print method to the Circuit class. The idea is
to give a visual representation of the quantum circuit.

The motivation for this is as a learning aid to help understand the
circuit created. The implementation can probably be made more efficient
but I wanted to get some feedback on this feature and if it is
worth pursuing/spending time on (or perhaps it could be useful as is
and developed further as required).