Add name to the gate type in qasm parser
This commit updates the gate type in grammar to include a name and
instead set type to 'gate'.

The motivation for this is a follow up commit where the type can be used
to collect all the gates from a parsed QASM file and the assemble it
into a qiskit-sim Circuit.