Make parser a member of class Parser
Currently parser is not a member of the Parser class and it looks like
the intention was to share it with all instances of class Parser. But it
also gets re-assigned in the constructor each time a new instance is
created. This commit makes parser a member instead, and perhaps a follow
up commit can take a look at sharing the parser (a quick test showed
that is was not as simple as I hoped).