add gate to gates map in constructor
This commit updates the Gate constructor to add the gate created to the
gates map if it does not exist.

The motivation for this was that when adding a custom gate it would not
be added to the gates map and hence an error would be thrown when trying
to execute the circuit.