allow custom gates to be overwritten
This commit adds the ability to overwrite/redefine the a custom gate.
The motivation for this came from wanting to have multiple circuits in a
program and allow them to have the same name for a custom gate, but the
matrix for the gate would be different for each circuit.