fix qiskit-cloud integration tests
This commit applies fixes to the cloud tests that connect to IBM Q by
running the test with the QK environment variable set:

QX_KEY=<API_TOKEN> npm t

There are still 3 failing tests which return a 400:

 3 failing

  1) algo:ibm:random
       should return a jobId:
     Error: GENERIC_ERROR
      at request (/qiskit-js/packages/qiskit-cloud/lib/request.js:45:13)
      at processTicksAndRejections (internal/process/task_queues.js:89:5)

  2) algo:ibm:result
       should return the result passing jobId:
     TypeError: The "jobId" parameter is mandatory (string)
      at Object.module.exports.result (index.js:75:11)
      at Context.t (test/functional.js:84:28)
      at processImmediate (internal/timers.js:439:21)

  3) algo:ibm:genBin
       should return a jobId:
     Error: GENERIC_ERROR
      at request (/qiskit-js/packages/qiskit-cloud/lib/request.js:45:13)
      at processTicksAndRejections (internal/process/task_queues.js:89:5)

Doing a search I found that this is a connection error but I'm able to
get a response from the server using a standalone example (taken from
the README.md) but I also get the same error with it. Opening this to
get some feedback and hopefully any pointer to what the issue might be.